// src/lib/types.ts
export interface BookInputs {
	pageWidth: number;
	pageHeight: number;
	blockThickness: number;
	boardThickness: number;
	coverMargin: number;
	fabricMargin: number;
}

export interface Dimensions {
	pageW: number;
	pageH: number;
	blockT: number;
	boardT: number;
	coverMargin: number;
	fabricMargin: number;
	boardW: number;
	boardH: number;
	spineW: number;
	gutter: number;
	fabricW: number;
	fabricH: number;
	cutOffset: number;
}